Add Makefile sections for manpages
For now this uses the output from `cabal --help` but the idea was to add
`cabal --full-help` to print the help messages for every subcommand.

But this won't be probably done because adding another global flag isn't
as easy as I previously thought and because the final format is still
not correct.

The correct `markdown` will be generated directly in the code.

This commit will most surely be drop when sending the proper PR to
upstream.